fixed extension check .cub
