+++ maps/invalid/.cub
@@ -1,7 +1,7 @@
-NO ./tiles/wall_n.png
-SO ./tiles/wall_s.png
-WE ./tiles/wall_w.png
-EA ./tiles/wall_e.png
+NO ./textures/ovi4.png
+SO ./textures/ovi3.png
+WE ./textures/ovi2.png
+EA ./textures/ovi1.png
 
 F 220,100,0
 C 225,226,0
