+++ maps/invalid/invalid_map_symbol.cub
@@ -9,5 +9,5 @@ C 225,226,0
 111111111111
 100000000001
 10000N000001
-1000000D0001
+1000000Y0001
 111111111111
