--- maps/invalid/map_not_closed_1.cub
@@ -1,16 +1,0 @@
-NO ./textures/north_texture.png
-SO ./textures/south_texture.png
-EA ./textures/east_texture.png
-WE ./textures/west_texture.png
-
-F 220,100,0
-C 225,30,0
-
-11111111
-1000000111111
-100000000000111
-100000011100001
-10N00001 100111
-10000001111 1
-1111100000001
-    111111111
