+++ maps/invalid/map_not_closed_3.cub
@@ -9,6 +9,6 @@ C 225,30,0
 11111111
 100S000111111
 100000000000111
-10101110111011
+101000111011
 10101 10 1101
 11111 111 111
