--- maps/invalid/over100enemy.cub
@@ -1,13 +1,0 @@
-NO ./tiles/wall/wall_n.png
-SO ./tiles/wall/wall_s.png
-WE ./tiles/wall/wall_w.png
-EA ./tiles/wall/wall_e.png
-
-F 44,42,74
-C 36,36,35
-
-1111111111111111111111111111111111111111
-1S000AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA1
-10000AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA1
-10000AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA1
-1111111111111111111111111111111111111111
