+++ maps/valid/map1.cub
@@ -6,7 +6,7 @@ F 00,00,00
 C 00,00,00
 
 111111111111111111
-10000000000011101
+1000000000001110D1
 100000000010000001
 11100N000000100001
 111111111111111111
